+++ src/components/ProductNotification/ProductNotification.tsx
@@ -14,6 +14,7 @@ const styles: StyleRulesCallback<ClassNames> = (theme: Theme) => ({
 interface Props {
   text: string;
   severity: 'minor' | 'major' | 'critical';
+  onClick?: () => void;
 }
 
 type CombinedProps = Props & WithStyles<ClassNames>;
@@ -21,7 +22,7 @@ type CombinedProps = Props & WithStyles<ClassNames>;
 const ProductNotifications: React.StatelessComponent<CombinedProps> = (props) => {
   const { text, severity } = props;
   const level = pathOr('warning', [severity], severityLevelMap);
-  return React.createElement(Notice, { flag: true, text, [level]: true });
+  return React.createElement(Notice, { flag: true, [level]: true }, text);
 };
 
 const styled = withStyles(styles, { withTheme: true });