+++ src/services/linodes/linodeActions.ts
@@ -74,3 +74,9 @@ export const startMutation = (linodeID: number) => {
   )
     .then(response => response.data)
 }
+
+export const scheduleOrQueueMigration = (linodeID: number) =>
+  Request<{}>(
+    setURL(`${API_ROOT}/linode/instances/${linodeID}/migrate`),
+    setMethod('POST'),
+  )