+++ CHANGELOG.md
@@ -30,6 +30,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Prevent user from submitting empty replies in support tickets
 
 
+
+## [0.35.0] - 2018-10.01
+### Added:
+  - Users can now provide feedback on Linode support in tickets.
+  - Added a welcome banner describing new features.
+  - Users can now migrate/mutate Linodes.
+
 ## [0.33.2] - 2018-09-11
 ### Added:
  - Profile